refactor(push): replace free-function helpers with collaborator classes
Replace the problematic free-function helper pattern with proper
constructor-injected collaborator classes:

- Add PushNotificationContentService for notification content extraction
- Add PushSessionEventReducer for event handling logic
- Add PushSessionMaintenanceService for maintenance operations
- Add PushSessionStateGraph for graph traversal operations
- Add PushSessionStateMutator for state mutations

Remove old helper files:
- push_notification_service_helpers.dart
- push_session_state_tracker_events.dart
- push_session_state_tracker_graph.dart
- push_session_state_tracker_maintenance.dart
- push_session_state_tracker_mutations.dart

Update tests to use new service instances instead of global functions.
Clean up unused imports in new collaborator files.

Follows AGENTS.md guidance: substantial logic belongs in named
classes with explicit dependencies, not free functions.
