+++ bridge/app/lib/src/push/push_maintenance_loop.dart
@@ -13,7 +13,6 @@ class PushMaintenanceLoop {
   final PushRateLimiter _rateLimiter;
   final Duration _maintenanceInterval;
   final int? Function() _rssBytesReader;
-  final void Function(String) _debugLogger;
   late final Timer _timer;
   PushMaintenanceTelemetrySnapshot? _lastSnapshot;
 
@@ -23,13 +22,11 @@ class PushMaintenanceLoop {
     required PushRateLimiter rateLimiter,
     Duration maintenanceInterval = const Duration(minutes: 10),
     int? Function()? rssBytesReader,
-    void Function(String)? debugLogger,
   }) : _tracker = tracker,
        _completionNotifier = completionNotifier,
        _rateLimiter = rateLimiter,
        _maintenanceInterval = maintenanceInterval,
-       _rssBytesReader = rssBytesReader ?? readCurrentRssBytes,
-       _debugLogger = debugLogger ?? Log.d {
+       _rssBytesReader = rssBytesReader ?? readCurrentRssBytes {
     _timer = Timer.periodic(_maintenanceInterval, (_) => runNow());
   }
 
@@ -66,7 +63,7 @@ class PushMaintenanceLoop {
           rssBytes: _rssBytesReader(),
         );
         _lastSnapshot = snapshot;
-        _debugLogger(snapshot.toLogMessage());
+        Log.d(snapshot.toLogMessage());
       },
     );
   }
